refactor: extract android install helpers
